backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.19.0",
    "numpy>=2.3.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.0",
]
